Added possibility for price as a Double

+++ app/src/main/java/pl/nataliana/mywine/ui/detail/AddWineFragment.kt
@@ -86,7 +86,7 @@ class AddWineFragment : Fragment() {
             data.getStringExtra(EXTRA_COLOR),
             data.getIntExtra(EXTRA_YEAR, 0),
             data.getFloatExtra(EXTRA_RATE, 0F),
-            data.getIntExtra(EXTRA_PRICE, 0),
+            data.getDoubleExtra(EXTRA_PRICE, 0.0),
             data.getStringExtra(EXTRA_TYPE)
         )
 
@@ -112,11 +112,11 @@ class AddWineFragment : Fragment() {
                     Integer.valueOf(0.toString())
                 }
             val rating: Float? = wineRating
-            val price: Int? =
+            val price: Double? =
                 try {
-                    Integer.valueOf(edit_text_price.text.toString())
+                    edit_text_price.text.toString().toDouble()
                 } catch (e: NumberFormatException) {
-                    Integer.valueOf(0.toString())
+                    0.toString().toDouble()
                 }
 
             putExtra(EXTRA_NAME, name)
